Add CO check to delete poll

+++ src/pages/api/deletePoll.ts
@@ -1,7 +1,11 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { authOptions } from '@/pages/api/auth/[...nextauth]';
 import { PrismaClient } from '@prisma/client';
 import * as Sentry from '@sentry/nextjs';
+import { getServerSession } from 'next-auth';
+
+import { checkIfCO } from '@/lib/checkIfCO';
 
 const prisma = new PrismaClient();
 
@@ -19,8 +23,24 @@ export default async function deletePoll(
   res: NextApiResponse<Data>,
 ): Promise<void> {
   const { pollId } = req.body;
-  // TODO: Add session check to verify it is coordinator. Also additional security step of verifying coordinator's signature before deleting poll
+  // TODO: Additional security step of verifying coordinator's signature before deleting poll
   try {
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      return res.status(401).json({
+        success: false,
+        message: 'User is not logged in',
+      });
+    }
+
+    const stakeAddress = session.user.stakeAddress;
+    const isCO = await checkIfCO(req, res, stakeAddress);
+    if (!isCO) {
+      return res.status(401).json({
+        success: false,
+        message: 'User is not a convention organizer',
+      });
+    }
     await prisma.poll.delete({
       where: {
         id: BigInt(pollId),
